+++ test/core/test_upnp_xml.cc
@@ -238,7 +238,7 @@ TEST_F(UpnpXmlTest, RenderObjectItemWithEscapes)
     auto clientConfig = std::make_shared<MyConfigMock>();
     auto clientManager = std::make_shared<ClientManager>(clientConfig, database);
     auto addr = std::make_shared<GrbNet>("192.168.99.100");
-    auto quirks = std::make_shared<Quirks>(subject, clientManager, addr, "DLNADOC/1.50");
+    auto quirks = std::make_shared<Quirks>(subject, clientManager, addr, "DLNADOC/1.50", nullptr);
 
     std::ostringstream expectedXml;
     expectedXml << "<DIDL-Lite>\n";
@@ -285,7 +285,7 @@ TEST_F(UpnpXmlTest, RenderObjectItemWithStrictXmlQuirks)
     auto clientConfig = std::make_shared<MyConfigMock>();
     auto clientManager = std::make_shared<ClientManager>(clientConfig, database);
     auto addr = std::make_shared<GrbNet>("192.168.99.100");
-    auto quirks = std::make_shared<Quirks>(subject, clientManager, addr, "Allegro-Software-WebClient/5.40b1 DLNADOC/1.50");
+    auto quirks = std::make_shared<Quirks>(subject, clientManager, addr, "Allegro-Software-WebClient/5.40b1 DLNADOC/1.50", nullptr);
 
     std::ostringstream expectedXml;
     expectedXml << "<DIDL-Lite>\n";
@@ -386,7 +386,7 @@ TEST_F(UpnpXmlTest, RenderObjectItemWithResourcesWithQuirks)
     ClientProfile pInfo;
     pInfo.flags = QUIRK_FLAG_CAPTION_PROTOCOL;
     auto addr = std::make_shared<GrbNet>("127.0.0.1");
-    struct ClientObservation client(addr, "ua", std::chrono::seconds(0), std::chrono::seconds(0), &pInfo);
+    struct ClientObservation client(addr, "ua", std::chrono::seconds(0), std::chrono::seconds(0), nullptr, &pInfo);
     auto quirks = std::make_shared<Quirks>(&client);
     // act
     subject->renderObject(obj, { "*" }, std::string::npos, root, quirks);