+++ ChangeLog.md
@@ -5,6 +5,7 @@
 - Add support for libsystemd
 - Add Test for formatSizeValue
 - Allow masking arguments
+- Check Headers in Quirks
 - CI Tests: Try more alpine versions
 - Command line option for even more config details
 - Documentation on Return Path Filtering